+++ test/jest/fixtures/index.js
@@ -12,3 +12,4 @@ export { address } from './address';
 export { exportReport } from './exportReport';
 export { hydratedOrder } from './hydratedOrder';
 export { instance } from './instance';
+export { integrationConfig } from './integrationConfig';