+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 * Add a return to Orders default search to app context menu dropdown. Refs UIOR-880.
 * Add two additional "Resource identifier" types to the product id type selection for POL. Refs UIOR-843.
 * Permission name discrepancies between `displayName` in `package.json` and translated values in `en_US.json`. Refs UIOR-734.
+* Allow user to set Acq method EDIFACT export triggers for POL. Refs UIOR-371.
 
 ## [3.0.3](https://github.com/folio-org/ui-orders/tree/v3.0.3) (2021-12-08)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v3.0.2...v3.0.3)